requirements.txt
beautifulsoup4==4.4.0
Django==1.8.3
django-oauth2-provider==0.2.6.1
requests==2.7.0
shortuuid==0.4.2
wheel==0.24.0
